Fix incorrect cell rendering. Refactor terminal code
Single cells in manual mode and adjacent fixed cells with the same value
were not displayed.

Refactored terminal related code. Moved helper functions into a separate
file. Moved private methods that don't access class' fields into a
separate file as functions
