+++ include/solver/utils.h
@@ -30,6 +30,8 @@ void fill_empty_board(Board &board);
 // calls solve() function for every adjacent cell
 bool solve_for_each_adjacent(Board &board, Region *p_region, Cell &cell);
 
+void concat_vectors(std::vector<Coord> &vect1, std::vector<Coord> &vect2); // concats vect2 into vect1
+
 template <class T> void reverse_vector(std::vector<T> &vect) {
     for (long unsigned int i = 0; i < vect.size() / 2; i++) {
         T temp = vect.at(i);
