+++ src/solver/solver.cpp
@@ -50,6 +50,7 @@ bool solve(Board &board, Region *p_region, Cell &cell) {
         int idx = get_next_unfilled_fixed_cell_idx(board);
         if (idx == -1) {
             board.result.push_back(cell.get_coord());
+            concat_vectors(board.result, adjs_with_same_val);
             return true;
         }
      
@@ -75,12 +76,14 @@ bool solve(Board &board, Region *p_region, Cell &cell) {
         } 
         
         board.result.push_back(cell.get_coord());
+        concat_vectors(board.result, adjs_with_same_val);
         return true;
     }
 
 
     if (solve_for_each_adjacent(board, p_region, cell)) {
         board.result.push_back(cell.get_coord());
+        concat_vectors(board.result, adjs_with_same_val);
         return true;
     }
 
@@ -94,6 +97,7 @@ bool solve(Board &board, Region *p_region, Cell &cell) {
  
         if (solve_for_each_adjacent(board, p_region, cell_in_reg)) {
             board.result.push_back(cell.get_coord());
+            concat_vectors(board.result, adjs_with_same_val);
             return true;
         }
     }
